Customer class update

+++ team-taskmanagement-core/src/main/java/com/team/taskmanagement/beans/CustomerDto.java
@@ -19,6 +19,6 @@ public class CustomerDto {
     private String zip;
     private String customerEmail;
     private String contact;
-    private int paymentTerm;
+    private short paymentTerm;
 
 }
